refactor: borrow logic
